Refactored EXIF functions
